requirements.txt
contourpy==1.3.2
cycler==0.12.1
et-xmlfile==2.0.0
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
nmrglue==0.11
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pip==25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
scipy==1.16.0
setuptools==78.1.1
six==1.17.0
tzdata==2025.2
uncertainties==3.2.3
wheel==0.45.1
xlrd==2.0.2
PySide6==6.9.1
seaborn==0.13.2